potted plant..
might be nicer without the pot? will see. will play with some more sprites too, probably.
heh. sprites.
billboarding things in the way that we want is going to be annoying. but we can surely do it. will need to relearn a lot of things.. probably best to just redo the code from scratch, honestly.